Initial move to VM architecture.
